requirements.txt
streamlit
transformers
torch
sentence-transformers
PyPDF2
python-docx
plotly
scikit-learn
